# Provide a faster way to get last update timestamp from Git log