Again, based on #401 so will need rebase after that's reviewed and fixed.

Fixes #406: the current implementation takes about 200ms for each file. That's 10.8s on a page like Italian (27 files x 2), even caching en-US results is not enough.

The idea is to get timestamps for all files calling `git log` only once, and cache the result. Even without cache (first load), the Italian page loads in less than 3 seconds. Second load goes back to 0.2s